+++ CHANGELOG.md
@@ -10,7 +10,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - Incident class accepts as a priority only a Priority enum
 - Incident class must be instantiated with a step number
-- Step number in Incident class must be an integer, null is not allowed anymore
+- Step number in Incident class is always an integer
+- Pool number in Incident class is always an integer
 
 ### Removed
 - Compatibility with PHP < 8.1